hw9 refactor: separated repo
