+++ currency-rates/internal/repo/subsctiption.go
@@ -11,10 +11,6 @@ type DBEmail struct {
 	IsActive bool
 }
 
-type DBUser struct {
-    Email string
-}
-
 func (repo *Repository) SetEmail(ctx context.Context, email string) error {
 	e := &DBEmail{
 		Email: email,
@@ -39,16 +35,4 @@ func (repo *Repository) DeactivateEmail(ctx context.Context, email string) error
 	}
 
 	return nil
-}
-
-func (repo *Repository) SetUser(ctx context.Context, email string) error {
-	u := &DBUser{
-		Email: email,
-	}
-
-	return repo.DB.Table("users").Create(u).Error
-}
-
-func (repo *Repository) DeleteUser(ctx context.Context, email string) error {
-	return repo.DB.Table("users").Where("email = ?", email).Delete(&DBUser{}).Error
 }
